requirements.txt
flickrapi
simplejson
requests
python-memcached